requirements.txt
asgiref==3.7.2
certifi==2023.11.17
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
daphne==4.0.0
dj-rest-auth==5.0.2
Django==4.2.9
django-allauth==0.54.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.6
packaging==23.2
pika==1.3.2
pillow==10.2.0
python3-openid==3.2.0
pytz==2023.3.post1
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
